+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ unreleased
 - [ADDED] estimation module, which provides state estimation functionality. The weighted least squares algorithm is used.
 - [FIXED] bug in voltage ratio for low voltage side tap changers
 - [ADDED] documentation of model validation and tests
+- [ADDED] DC power flow function pp.rundcopp
 
 [1.0.2] - 2016-11-30
 ----------------------